+++ src/commands/commandManager.ts
@@ -13,8 +13,12 @@ export class CommandManager {
     this.variablePanelManager = variablePanelManager;
   }
 
+  private async getCustomVariables(): Promise<{ [key: string]: string[] }> {
+    const config = vscode.workspace.getConfiguration("jinja2-html-enhancer");
+    return config.get('customVariables', {});
+  }
 
-  public checkVariables() {
+  public async checkVariables() {
     vscode.window.showInformationMessage(I18n.__('variable.checkingVariables'));
     const editor = vscode.window.activeTextEditor;
     if (editor) {
@@ -41,13 +45,18 @@ export class CommandManager {
 
   public async saveVariable(diagnosticMessage: string) {
     const variable = extractVariableName(diagnosticMessage);
+    const activeEditor = vscode.window.activeTextEditor;
 
     if (!variable) {
       vscode.window.showWarningMessage(I18n.__('error.variableNotFound'));
       return;
     }
 
-    const config = vscode.workspace.getConfiguration("jinja2-html-enhancer");
+    if (!activeEditor) {
+      vscode.window.showWarningMessage(I18n.__('error.noActiveEditor'));
+      return;
+    }
+    const filePath = activeEditor.document.uri.fsPath;
 
     const target = vscode.workspace.workspaceFolders
       ? vscode.ConfigurationTarget.WorkspaceFolder
@@ -57,9 +66,19 @@ export class CommandManager {
       ? I18n.__('quickFix.workspaceTarget')
       : I18n.__('quickFix.globalTarget');
 
-    const currentVariables = config.get<{ [key: string]: string }>('customVariables') || {};
-    currentVariables[variable] = variable;
+    const config = vscode.workspace.getConfiguration("jinja2-html-enhancer");
 
+    const currentVariables: { [key: string]: string[] } = config.get('customVariables', {});
+
+    if (currentVariables[filePath] && currentVariables[filePath].includes(variable)) {
+      vscode.window.showWarningMessage(I18n.__('variable.variableExists', { variable }));
+      return;
+    }
+
+    if (!currentVariables[filePath]) {
+      currentVariables[filePath] = [];
+    }
+    currentVariables[filePath].push(variable);
     try {
       await config.update('customVariables', currentVariables, target);
       vscode.window.showInformationMessage(I18n.__('variable.variableSaved', { variable, targetTranslation }));